Refactored bookmark related code out of ReviewResource
1. refs #336
