+++ hoot-services/src/main/java/hoot/services/writers/review/ReviewBookmarksRemover.java
@@ -38,7 +38,8 @@ import org.slf4j.LoggerFactory;
 import com.mysema.query.sql.Configuration;
 import com.mysema.query.sql.dml.SQLDeleteClause;
 
-public class ReviewBookmarksRemover {
+public class ReviewBookmarksRemover 
+{
 	@SuppressWarnings("unused")
   private static final Logger log = LoggerFactory.getLogger(ReviewBookmarksRemover.class);
 	
@@ -46,7 +47,7 @@ public class ReviewBookmarksRemover {
 	
 	public ReviewBookmarksRemover(final Connection cn) 
 	{
-		_conn = cn;
+	  _conn = cn;
 	}
 	
 	/**
@@ -59,16 +60,8 @@ public class ReviewBookmarksRemover {
 	public long remove(final ReviewBookmarkDelRequest request) throws Exception
 	{
 		long nDeleted = 0;
-		
-		try
-		{
-			SQLDeleteClause cl = _createDelClause(request);
-			nDeleted = cl.execute();
-		}
-		catch (Exception ex)
-		{
-			
-		}
+		SQLDeleteClause cl = _createDelClause(request);
+		nDeleted = cl.execute();
 		return nDeleted;
 	}
 	
@@ -83,19 +76,10 @@ public class ReviewBookmarksRemover {
 	protected SQLDeleteClause _createDelClause(final ReviewBookmarkDelRequest request) throws Exception
 	{
 		SQLDeleteClause cl = null;
-		
-		try
-		{
-			QReviewBookmarks reviewBookmarks = QReviewBookmarks.reviewBookmarks;
-			Configuration configuration = DbUtils.getConfiguration();
-			cl = new SQLDeleteClause(_conn, configuration, reviewBookmarks)
-			.where(reviewBookmarks.id.eq(request.getBookmarkId()));
-		}
-		catch (Exception ex)
-		{
-			
-		}
-		
+		QReviewBookmarks reviewBookmarks = QReviewBookmarks.reviewBookmarks;
+		Configuration configuration = DbUtils.getConfiguration();
+		cl = new SQLDeleteClause(_conn, configuration, reviewBookmarks)
+		.where(reviewBookmarks.id.eq(request.getBookmarkId()));
 		return cl;
 	}
 }
